Move function related struct to a separate file
this commit move a handful of function related struct to a separate file while changing as little as possible.
 It only change the visibility of certain element but it do not attempt to untangle them.
